add interfaces to set message header size and handle packages

+++ src/packages/packet_helper.cpp
@@ -3,7 +3,7 @@
 
 namespace tsp_client {
 
-    uint32_t PackHelper::get_message_header_size() {
+    uint8_t PackHelper::get_message_header_size() {
         return sizeof(MessageHeader);
     }
